refactor(stores): extract persist migrations to migrations.ts
Plain data-in/data-out fns the store configs call.
