+++ src/stores/ui.ts
@@ -2,30 +2,17 @@ import { create } from "zustand";
 import { persist } from "zustand/middleware";
 import { applyZoom, clampZoom, DEFAULT_ZOOM } from "../lib/zoom.ts";
 import { applyDensity, normalizeDensity, DEFAULT_DENSITY, type Density } from "../lib/density.ts";
+import {
+  DEFAULT_EXPORT_METHODS,
+  DEFAULT_SIDEBAR_PAGES,
+  migrateUi,
+  type ExportFormatKey,
+  type ExportMethods,
+  type SidebarPageKey,
+  type SidebarPages,
+} from "./migrations.ts";
 
-export type SidebarPageKey = "graph" | "search" | "doi" | "tags" | "notes" | "shared" | "reading";
-
-export type SidebarPages = Record<SidebarPageKey, boolean>;
-
-const DEFAULT_SIDEBAR_PAGES: SidebarPages = {
-  graph: true,
-  search: true,
-  doi: true,
-  tags: false,
-  notes: false,
-  shared: true,
-  reading: true,
-};
-
-export type ExportFormatKey = "lxproj" | "bibtex" | "obsidian";
-
-export type ExportMethods = Record<ExportFormatKey, boolean>;
-
-const DEFAULT_EXPORT_METHODS: ExportMethods = {
-  lxproj: true,
-  bibtex: true,
-  obsidian: true,
-};
+export type { ExportFormatKey, ExportMethods, SidebarPageKey, SidebarPages };
 
 interface UiState {
   sidebarCollapsed: boolean;
@@ -87,33 +74,7 @@ export const useUiStore = create<UiState>()(
     {
       name: "linxiv-ui",
       version: 7,
-      migrate(persisted, fromVersion) {
-        const state = (persisted ?? {}) as Partial<UiState>;
-        if (fromVersion < 1) {
-          state.sidebarPages = { ...DEFAULT_SIDEBAR_PAGES, ...state.sidebarPages };
-        }
-        if (fromVersion < 2) {
-          state.exportMethods = { ...DEFAULT_EXPORT_METHODS, ...state.exportMethods };
-        }
-        if (fromVersion < 3) {
-          state.zoom = DEFAULT_ZOOM;
-        }
-        if (fromVersion < 4) {
-          state.hideSingleAuthors = false;
-        }
-        if (fromVersion < 5) {
-          state.density = DEFAULT_DENSITY;
-        }
-        if (fromVersion < 6) {
-          // Backfill the new "shared" page key into persisted sidebarPages.
-          state.sidebarPages = { ...DEFAULT_SIDEBAR_PAGES, ...state.sidebarPages };
-        }
-        if (fromVersion < 7) {
-          // Backfill the new "reading" page key into persisted sidebarPages.
-          state.sidebarPages = { ...DEFAULT_SIDEBAR_PAGES, ...state.sidebarPages };
-        }
-        return state;
-      },
+      migrate: migrateUi,
       // The webview starts every launch at 100%; re-apply the saved zoom once
       // the persisted value is loaded (and normalize it in case the stored
       // number is out of range or corrupt).
